5g-network-monitoring/requirements.txt
pandas
numpy
tensorflow
tensorflow-privacy
scikit-learn
requests
langgraph
joblib
beautifulsoup4
geopy
python-eventbrite
flask
flask-login
requests-oauthlib
cryptography
kafka-python
shap
osmnx
tweepy
cachetools